backend/requirements.txt
Django==3.2
djangorestframework
django-filter
pytest
pytest-django
